requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
tzdata==2023.3
djangorestframework==3.14.0
model-bakery==1.10.1
Faker==18.3.1
pytest-django==4.2.0
requests==2.28.2
factory-boy==3.2.1
django-debug-toolbar==3.8.1
